+++ galsim/base.py
@@ -1814,4 +1814,3 @@ class DeVaucouleurs(GSObject):
         """
         return self.SBProfile.getScaleRadius()
 
-