+++ galsim/config/gsobject.py
@@ -32,6 +32,7 @@ valid_gsobject_types = {
     'Ring' : '_BuildRing',
     'Pixel' : '_BuildPixel',
     'RealGalaxy' : '_BuildRealGalaxy',
+    'RealGalaxyOriginal' : '_BuildRealGalaxyOriginal'
 }
 
 class SkipThisObject(Exception):
@@ -433,7 +434,7 @@ def _BuildRealGalaxy(config, key, base, ignore, gsparams):
         index = kwargs['index']
         if index >= real_cat.nobjects:
             raise IndexError(
-                "%s index has gone past the number of entries in the catalog"%param_name)
+                "%s index has gone past the number of entries in the catalog"%index)
 
     gal = galsim.RealGalaxy(real_cat, **kwargs)
 
@@ -443,6 +444,13 @@ def _BuildRealGalaxy(config, key, base, ignore, gsparams):
     if not whiten: del gal.noise
 
     return gal, safe
+
+
+def _BuildRealGalaxyOriginal(config, key, base, ignore, gsparams):
+    """@brief Return the original image from a RealGalaxy instance defined by user input.
+    """
+    image, safe = _BuildRealGalaxy(config, key, base, ignore, gsparams)
+    return image.original_image, safe    
 
 
 def _BuildSimple(config, key, base, ignore, gsparams={}):