+++ include/galsim/SBInterpolatedImageImpl.h
@@ -39,8 +39,10 @@ namespace galsim {
             double dx, double pad_factor, const GSParamsPtr& gsparams);
 
         SBInterpolatedImageImpl(
-            const MultipleImageHelper& multi, const std::vector<double>& weights,
-            boost::shared_ptr<Interpolant2d> xInterp, boost::shared_ptr<Interpolant2d> kInterp,
+            const MultipleImageHelper& multi,
+            const std::vector<double>& weights,
+            boost::shared_ptr<Interpolant2d> xInterp,
+            boost::shared_ptr<Interpolant2d> kInterp,
             const GSParamsPtr& gsparams);
 
         ~SBInterpolatedImageImpl();